Move dev-only methods to a partial class file
